added builtin commands handler stub & changed method of storing process ids
- cush.c

stubbed out the method that will process built in commands. Moved the code from the main that proccessed the built in commands to
the function

removed the _add_pid_to_job function as we changed our approach of storing pids. Now the pids will be stored in each command struct
to aviod having to deal with memory allocations

- shell-ast.c
added pid variable tosreo process pid in the ast_command struct