added function that searches for a job given a prcess id
Hey Hey
In this commit the major difference is I remade the function that used to search for a job  using group id to
it searching for job using a process id. The problem with searching using the gpid given the pid is that if the process termiantes
before the signal handler gets to it, it won't be able to get pid from the gpid. So instead we stored the pid of each process.
when the signal handler recieves a status change of a certain process it uses that to identify which jon that process belongs to.

Next steps are to move some code in different file for cleanliness and readability and to manage deleting jobs when all their processes
exit.