Added support for input & output io redirection
The main change in this commit is the set_up_io function. The function allows user to use <, > io redirection.
This was done using posix_spawn_file_actions_addopen with different configuration depending on the pass flag (__ioflag)

__ioflag:
	- IO_IN
		- used  for using a file as input to a command
		- sets up STDIN to map to that file
	- IO_OUT
		- used for using a file as output from a command
		sets up STDOUT to map to that file
Issues:
One issues that is prevelent is when running a command like "cat t > test" assuming t is a nonexistant file.
since the output of cat is redirected into test when the cat process realizes "t" doesn't exist the output error is sored in the file "test". The provided test are passing. However, it's an issue to be looked at.

Next up piping